requirements.txt
umap-learn
numpy
keras
matplotlib
seaborn
scikit-learn
tqdm
kymatio
torch
torchvision
hdbscan
pandas
joblib
pynndescent
imblearn
tensorflow